requirements.txt
flask==3.0.3
gunicorn==21.2.0
kafka-python==2.0.2
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
requests==2.32.3
prometheus-client==0.20.0
